Separate out and uncomment code
