+++ src/applications/covid19-chatbot/createCoronavirusChatbot.js
@@ -1,0 +1,15 @@
+export default (_store, widgetType) => {
+  // Derive the element to render our widget.
+  const root = document.querySelector(`[data-widget-type="${widgetType}"]`);
+
+  // Escape early if no element was found.
+  if (!root) {
+    return;
+  }
+
+  // webpackChunkName: "find-va-forms"
+  import('./index').then(module => {
+    const initializeChatbot = module.default;
+    initializeChatbot();
+  });
+};
